Add script for building and starting the application with gradle
While developing, it useful to have a script that (re)build the application
and run it. The gradle run command do exactly this:

./gradlew run

The bad thing is, that command line arguments cannot be passed that simple
with the run command. So a little trick is used to pass the arguments to
the application, now the following command can be executed to rebuild the
application and run it with arguments:

./scripts/wifi-qrcode-generator-dev.sh --ssid "foo" -p "bar" -a WPA example.png